+++ frontend/packages/erd-core/src/features/erd/components/ERDContent/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useApplyUrlState'
 export * from './useGroupMembership'
 export * from './useInitialAutoLayout'
 export * from './useQueryParamsChanged'